+++ src/mailmanclient/docs/NEWS.rst
@@ -11,6 +11,7 @@ NEWS for mailmanclient
   ``MailingList.unsubscribe``. (Fixes #62)
 - Add support to fetch pending unsubscription requests. (Closes #63)
 - Add ``member_id`` as a property of ``Member`` object. (Closes #64)
+- Return pending token when a Member is unsubscribed. (Closes #65)
 
 .. _news-3.3.2: